Refactor playwright login helper
It now works for all test specs and does not require a password from the
environment.
