+++ php/tests/tests/initial-setup.spec.js
@@ -1,21 +1,11 @@
 import { test, expect } from '@playwright/test';
 import { writeFileSync } from 'node:fs'
+import { logInToContainersPage } from './helpers.js';
 
 test('Initial setup', async ({ page: setupPage }) => {
   test.setTimeout(10 * 60 * 1000)
 
-  // Extract initial password
-  await setupPage.goto('./setup');
-  const password = await setupPage.locator('#initial-password').innerText()
-  const containersPagePromise = setupPage.waitForEvent('popup');
-  await setupPage.getByRole('link', { name: 'Open Nextcloud AIO login ↗' }).click();
-  const containersPage = await containersPagePromise;
-
-  // Log in and wait for redirect
-  await containersPage.locator('#master-password').click();
-  await containersPage.locator('#master-password').fill(password);
-  await containersPage.getByRole('button', { name: 'Log in' }).click();
-  await containersPage.waitForURL('./containers');
+  const containersPage = await logInToContainersPage(setupPage);
 
   // Reject IP addresses
   await containersPage.locator('#domain').click();
